requirements.txt
# Core API & Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=0.20.0 # Relaxed constraint, was pinned for TorBot
pydantic>=2.0.0

# NLP/LLM Tools
openai>=1.0.0
langchain>=0.1.0
# unstructured>=0.10.0 # Remains commented out
# lxml==5.1.0 # Remains commented out

# Add requests library
requests

# Crawler & Task Queue
botasaurus==3.2.5 # Pin to a specific 3.x version, used by the worker
rq>=1.15.0
redis>=5.0.0
beautifulsoup4 # Added for parsing in @request scraper
supabase>=1.0.0 # For ThreatParser (adjust version as needed, e.g., specific 1.x or 2.x)
pydantic-settings>=2.0.0 # For seer.utils.config used by ThreatParser